requirements.txt
aiohttp>=3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
deepgram-sdk==0.2.4
django==4.0.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
frozenlist==1.3.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
multidict==6.0.2
nltk==3.8.1
oauthlib==3.2.0
openai==0.26.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-dotenv==0.19.2
requests==2.27.1
requests-oauthlib==1.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.1.0
txaio==22.2.1
typing_extensions==4.1.1
urllib3==1.26.8
websockets==10.2
wikipedia-api==0.5.8
yarl==1.7.2
zope.interface==5.4.0
